move classproperty to its own file
