+++ pennylane/_class_property.py
