+++ pennylane/estimator/resource_operator.py
@@ -23,7 +23,7 @@ from typing import Any
 import numpy as np
 
 import pennylane.estimator.ops as qre_ops
-from pennylane.operation import classproperty
+from pennylane._class_property import classproperty
 from pennylane.queuing import QueuingManager
 from pennylane.typing import TensorLike
 from pennylane.wires import Wires
